custom_gpt/requirements.txt
acres==0.3.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
async-timeout==5.0.1
attr==0.3.1
attrs==25.3.0
azure-core==1.35.0
azure-storage-blob==12.26.0
black==25.1.0
bleach==5.0.1
boto==2.49.0
boto3==1.40.21
botocore==1.40.21
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.2
datamodel-code-generator==0.26.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.11
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
dotenv==0.9.9
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email-validator==2.3.0
etelemetry==0.3.1
exceptiongroup==1.2.2
expiringdict==1.2.2
Faker==37.6.0
filelock==3.18.0
genson==1.3.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==5.6.2
inflection==0.5.1
isodate==0.6.1
isort==5.13.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
label-studio==1.20.0
label-studio-sdk==1.0.18
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
looseversion==1.3.0
lxml==5.3.2
lxml_html_clean==0.4.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy_extensions==1.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==1.26.4
openai==1.70.0
opencv-python==4.11.0.86
opentelemetry-api==1.36.0
ordered-set==4.0.2
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.4.0
proto-plus==1.26.1
protobuf==6.32.0
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydot==3.0.4
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
pyRFC3339==2.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-json-logger==2.0.4
pytz==2022.7.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
redis==5.2.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
rich==14.1.0
rpds-py==0.27.1
rq==1.16.2
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.13.1
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentry-sdk==2.35.2
simplejson==3.20.1
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tldextract==5.3.0
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.11.0
uritemplate==4.2.0
urllib3==1.26.20
user-agents==2.2.0
webencodings==0.5.1
wrapt==1.17.3
xmljson==0.2.1
zipp==3.23.0
